GH-980 change the way we update push token
