+++ MultisigWallet/MultisigWalletApplication/WalletApplicationServiceTests/BaseWalletApplicationServiceTests.swift
@@ -18,7 +18,6 @@ class BaseWalletApplicationServiceTests: XCTestCase {
     let ethereumService = MockEthereumApplicationService()
     let transactionService = TransactionDomainService()
     let notificationService = MockNotificationService()
-    let tokensService = MockPushTokensDomainService()
     let transactionRepository = InMemoryTransactionRepository()
     let relayService = MockTransactionRelayService(averageDelay: 0, maxDeviation: 0)
     let encryptionService = MockEncryptionService()
@@ -56,7 +55,6 @@ class BaseWalletApplicationServiceTests: XCTestCase {
         DomainRegistry.put(service: portfolioRepository, for: SinglePortfolioRepository.self)
         DomainRegistry.put(service: accountRepository, for: AccountRepository.self)
         DomainRegistry.put(service: notificationService, for: NotificationDomainService.self)
-        DomainRegistry.put(service: tokensService, for: PushTokensDomainService.self)
         DomainRegistry.put(service: relayService, for: TransactionRelayDomainService.self)
         DomainRegistry.put(service: tokenItemsRepository, for: TokenListItemRepository.self)
         DomainRegistry.put(service: tokenItemsService, for: TokenListDomainService.self)
